refactor avatar
