+++ client/src/components/AppBar/AppBar.css
@@ -1,5 +1,3 @@
 .colorApp {
     background-color: #113;
-    /* background-color: black; */
-    
 }
