+++ client/src/components/AppBar/AppBar.jsx
@@ -1,3 +1,5 @@
+import './AppBar.css';
+
 import * as React from 'react';
 import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
@@ -7,9 +9,9 @@ import Typography from '@mui/material/Typography';
 import Menu from '@mui/material/Menu';
 import MenuIcon from '@mui/icons-material/Menu';
 import Container from '@mui/material/Container';
-import Avatar from '@mui/material/Avatar';
+// import Avatar from '@mui/material/Avatar';
 import Button from '@mui/material/Button';
-import Tooltip from '@mui/material/Tooltip';
+// import Tooltip from '@mui/material/Tooltip';
 import MenuItem from '@mui/material/MenuItem';
 // import AdbIcon from '@mui/icons-material/Adb';
 // import NotificationsIcon from '@mui/icons-material/Notifications';
@@ -17,8 +19,7 @@ import MenuItem from '@mui/material/MenuItem';
 import Stack from '@mui/material/Stack';
 import NotificationBell from './NotificationBell';
 import Logo from './Logo';
-import './AppBar.css';
-import AvatarIcon from './AvatarIcon';
+import AvatarIcon from '../AvatarIcon';
 
 const pages = ['Meus Projetos', 'Descobrir'];
 const settings = ['Perfil', 'Sair'];
