+++ client/src/components/AvatarIcon/index.jsx
